v17: Add support for network methods
Add support to the `v17` client and integration tests for network
section methods.

Note, we do not implement `into_model` for the network methods because
apart from `FeeRate` there are no `rust-bitcoin` types required.

Note also `getaddednodeinfo` is untested because without configuring the
network it returns an empty list so a test within the current framework
would be smoke and mirrors.