v17: Update key
Update the key docs now that we worked out that a bunch of RPC methods
are either deprecated or do not return anything.

We choose not to implement deprecated for v17 because v17.1 is so far
back anyways. And obviously if nothing is returned we don't need a type
for it.