Move macro call
During development of the v28 support we re-ordered some macro calls,
so as not to hold up an external contributor with triviality I acked
and merged as it was.

Put the macro calls in the same order so that when one diffs
`v27_api.rs` and `v28_api.rs` there are no false positives in the
diff.

Internal change only.